BDD: big refactor on BrowserContext to simplify maintenance
