+++ src/EzSystems/BehatBundle/Features/Context/Browser/SubContexts/BrowserSubContext.php
@@ -1,0 +1,22 @@
+<?php
+/**
+ * File containing the BrowserSubContext class.
+ *
+ * This is the parent object of all Browser sub contexts
+ *
+ * @copyright Copyright (C) 1999-2014 eZ Systems AS. All rights reserved.
+ * @license http://www.gnu.org/licenses/gpl-2.0.txt GNU General Public License v2
+ * @version //autogentag//
+ */
+
+namespace EzSystems\BehatBundle\Features\Context\Browser\SubContexts;
+
+use Behat\Behat\Context\BehatContext;
+
+/**
+ * BrowserSubContext
+ */
+abstract class BrowserSubContext extends BehatContext
+{
+
+}
